Opt Staff #14

+++ AttendanceManagement/Views/Staff.xaml.cs
@@ -28,10 +28,10 @@ namespace AttendanceManagement.Views
 
 
 
-        SqlConnection conn = new SqlConnection(@"Data Source=ADAM-DELL;Initial Catalog=AttendanceManagement;Integrated Security=True");
+        SqlConnection conn = new SqlConnection(@"Data Source=DESKTOP-QUDI77S\MSSQLSERVER01;Initial Catalog=AttendanceManagement;Integrated Security=True");
         SqlCommand Cmd;
-        //SqlDataAdapter Sda;
-        //SqlDataReader dr;
+        SqlDataAdapter Sda;
+        SqlDataReader dr;
         DataSet ds = new DataSet();
         DataTable dt = new DataTable();
 
@@ -39,16 +39,34 @@ namespace AttendanceManagement.Views
 
 
 
+        public void remplir_combo()
+        {
+
+            Cmd = new SqlCommand("select [Class Name] from Classes", conn);
+            SqlDataReader dr = Cmd.ExecuteReader();
+            while (dr.Read())
+            {
+                comb_classes.Items.Add(dr["Class Name"]);
+            }
+            dr.Close();
+        }
+
 
         DataTable Dt = new DataTable();
 
         private void Grid_Loaded(object sender, RoutedEventArgs e)
         {
+            comb_classes.Text = "c#";
+            string combo_text = comb_classes.Text;
+
+
             dg.Visibility = Visibility.Visible;
             dg1.Visibility = Visibility.Hidden;
             conn.Open();
 
-            Cmd = new SqlCommand("select u.[User Id],u.[Full Name],u.Email,c.[Class Name] from Users u inner join Classes c on u.[Class Id]=c.[Id Class] where u.[Role Id]=4", conn);
+            remplir_combo();
+
+            Cmd = new SqlCommand("select u.[User Id],u.[Full Name],u.Email,c.[Class Name] from Users u inner join Classes c on u.[Class Id]=c.[Id Class] where u.[Role Id]=4 and c.[Class Name]='" + combo_text + "'", conn);
             SqlDataReader dr = Cmd.ExecuteReader();
             DataTable t = new DataTable();
             t.Load(dr);
@@ -78,10 +96,13 @@ namespace AttendanceManagement.Views
 
             DataRowView row = dg.SelectedItem as DataRowView;
             int id_student = Convert.ToInt32(row.Row[0].ToString());
-            conn.Open();
-            SqlCommand cmd1 = new SqlCommand("delete from Attendance where [Student id]='" + id_student + "' and Date = '" + date + "' and Absent = 'oui'", conn);
-            cmd1.ExecuteNonQuery();
-            conn.Close();
+            if (conn.State == ConnectionState.Closed)
+            {
+                conn.Open();
+                SqlCommand cmd1 = new SqlCommand("delete from Attendance where [Student id]='" + id_student + "' and Date = '" + date + "' and Absent = 'oui'", conn);
+                cmd1.ExecuteNonQuery();
+                conn.Close();
+            }
 
         }
 
@@ -102,15 +123,18 @@ namespace AttendanceManagement.Views
         {
             DataRowView row = dg.SelectedItem as DataRowView;
             int id_student = Convert.ToInt32(row.Row[0].ToString());
-            conn.Open();
-            SqlCommand cmd1 = new SqlCommand("delete from Attendance where [Student id]='" + id_student + "' and Date = '" + date + "'and retard = 'oui'", conn);
-            cmd1.ExecuteNonQuery();
-            conn.Close();
+            if (conn.State == ConnectionState.Closed)
+            {
+                conn.Open();
+                SqlCommand cmd1 = new SqlCommand("delete from Attendance where [Student id]='" + id_student + "' and Date = '" + date + "'and retard = 'oui'", conn);
+                cmd1.ExecuteNonQuery();
+                conn.Close();
+            }
         }
 
         private void afficher_les_absents_Click(object sender, RoutedEventArgs e)
         {
-
+            string combo_text = comb_classes.Text;
             if (conn.State == ConnectionState.Closed)
             {
                 conn.Open();
@@ -118,7 +142,7 @@ namespace AttendanceManagement.Views
                 dg1.Visibility = Visibility.Visible;
 
 
-                Cmd = new SqlCommand("select a.[Date] , u.[Full Name] , a.[Description] ,c.[Class Name] From Users u inner join Attendance a on a.[Student Id] =u.[User Id] inner join Classes c on c.[Id Class]=u.[Class Id]", conn);
+                Cmd = new SqlCommand("select a.[Date] , u.[Full Name] , a.[Description] ,c.[Class Name] From Users u inner join Attendance a on a.[Student Id] =u.[User Id] inner join Classes c on c.[Id Class]=u.[Class Id] where a.[Date]='" + date + "' and c.[Class Name]='" + combo_text + "' ", conn);
                 SqlDataReader dr = Cmd.ExecuteReader();
                 DataTable t = new DataTable();
                 t.Load(dr);
@@ -131,7 +155,7 @@ namespace AttendanceManagement.Views
                 dg1.Visibility = Visibility.Visible;
 
 
-                Cmd = new SqlCommand("select u.[User Id],u.[Full Name],u.Email,c.[Class Name] from Users u inner join Classes c on u.[Class Id]=c.[Id Class] where u.[Role Id]=4", conn);
+                Cmd = new SqlCommand("select a.[Date], u.[Full Name], a.[Description], c.[Class Name] From Users u inner join Attendance a on a.[Student Id] = u.[User Id] inner join Classes c on c.[Id Class] = u.[Class Id] where a.[Date]='" + date + "' and c.[Class Name]='" + combo_text + "' ", conn);
                 SqlDataReader dr = Cmd.ExecuteReader();
                 DataTable t = new DataTable();
                 t.Load(dr);
@@ -154,10 +178,13 @@ namespace AttendanceManagement.Views
             check_retard_unchecked(sender, e);
             DataRowView row = dg.SelectedItem as DataRowView;
             int id_student = Convert.ToInt32(row.Row[0].ToString());
-            conn.Open();
-            SqlCommand cmd1 = new SqlCommand("INSERT INTO Attendance (Date, IsJustified, [Student id],Absent,Retard) VALUES ('" + date + "','false','" + id_student + "', 'oui' , 'non')", conn);
-            cmd1.ExecuteNonQuery();
-            conn.Close();
+            if (conn.State == ConnectionState.Closed)
+            {
+                conn.Open();
+                SqlCommand cmd1 = new SqlCommand("INSERT INTO Attendance (Date, IsJustified, [Student id],Absent,Retard) VALUES ('" + date + "','false','" + id_student + "', 'oui' , 'non')", conn);
+                cmd1.ExecuteNonQuery();
+                conn.Close();
+            }
         }
 
         private void text_search_TextChanged(object sender, TextChangedEventArgs e)
@@ -198,6 +225,42 @@ namespace AttendanceManagement.Views
 
 
 
+        }
+
+        private void comb_classes_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            string combo_text = comb_classes.SelectedValue.ToString();
+            MessageBox.Show(combo_text);
+            if (conn.State == ConnectionState.Closed)
+            {
+                conn.Open();
+
+                dg.Visibility = Visibility.Hidden;
+                dg1.Visibility = Visibility.Visible;
+
+
+                Cmd = new SqlCommand("select u.[User Id],u.[Full Name],u.Email,c.[Class Name] from Users u inner join Classes c on u.[Class Id]=c.[Id Class] where u.[Role Id]=4 and c.[Class Name]='" + combo_text + "'", conn);
+                SqlDataReader dr = Cmd.ExecuteReader();
+                DataTable t = new DataTable();
+                t.Load(dr);
+                dr.Close();
+                dg1.ItemsSource = t.DefaultView;
+            }
+            else
+            {
+                dg.Visibility = Visibility.Hidden;
+                dg1.Visibility = Visibility.Visible;
+
+
+
+
+                Cmd = new SqlCommand("select u.[User Id],u.[Full Name],u.Email,c.[Class Name] from Users u inner join Classes c on u.[Class Id]=c.[Id Class] where u.[Role Id]=4 and c.[Class Name]='" + combo_text + "'", conn);
+                SqlDataReader dr = Cmd.ExecuteReader();
+                DataTable t = new DataTable();
+                t.Load(dr);
+                dr.Close();
+                dg1.ItemsSource = t.DefaultView;
+            }
         }
     }
 }
